fix: Clear clippy under --all-targets --all-features
The default lint command (just lint) only checks default features, so the
object_id and chrono feature code was never linted. Fix all clippy errors
surfaced under --all-features on both beta and stable:

- Split flat inline test modules into the wrapper pattern so test fns are
  recognized as in-test (separate #[cfg(test)] + #[cfg(feature)] attrs)
- str_to_string, default_numeric_fallback, arbitrary_source_item_ordering,
  doc_markdown, absolute_paths, shadow_reuse, unused_trait_names
- Allow missing_trait_methods (unsatisfiable for Drop / serde Visitor)
