+++ src/field_type.rs
@@ -6,6 +6,11 @@ use syn::Attribute;
 use crate::features::model_schema_prop::ModelSchemaPropMeta;
 use crate::utils::{lookup_alias_info, safe_type_name};
 
+#[cfg(feature = "chrono")]
+use crate::features::chrono;
+#[cfg(feature = "object_id")]
+use crate::features::object_id;
+
 #[cfg(feature = "serde")]
 use crate::features::serde::{
     parse_serde_field_attributes as parse_serde_field_attributes_impl,
@@ -56,11 +61,11 @@ pub enum FieldDefType {
     /// Boolean primitive - maps to boolean.
     Boolean,
     #[cfg(feature = "chrono")]
-    /// Chrono `DateTime<Tz>` type - requires "`chrono`" feature
-    /// Maps to `string` in TS (ISO 8601 format: "2025-11-29T14:30:00Z")
-    /// Zod: z.string().datetime()
-    /// JSON Schema: string with format "date-time"
-    /// Note: The timezone type parameter is ignored for schema generation
+    /// Chrono `DateTime<Tz>` type - requires "`chrono`" feature.
+    /// Maps to `string` in TS (ISO 8601 format: "2025-11-29T14:30:00Z").
+    /// Zod: `z.string().datetime()`.
+    /// JSON Schema: string with format "date-time".
+    /// Note: the timezone type parameter is ignored for schema generation.
     DateTime,
     F32,
     F64,
@@ -73,29 +78,29 @@ pub enum FieldDefType {
     /// Boxed for recursion. Generates Partial<Record<K, V>> in TS.
     Map(Box<FieldDef>, Box<FieldDef>),
     #[cfg(feature = "chrono")]
-    /// Chrono `NaiveDate` type - requires "`chrono`" feature
-    /// Maps to `string` in TS (ISO 8601 date format: "2025-11-29")
-    /// Zod: z.string().date()
-    /// JSON Schema: string with format "date"
+    /// Chrono `NaiveDate` type - requires "`chrono`" feature.
+    /// Maps to `string` in TS (ISO 8601 date format: "2025-11-29").
+    /// Zod: `z.string().date()`.
+    /// JSON Schema: string with format "date".
     NaiveDate,
     #[cfg(feature = "chrono")]
-    /// Chrono `NaiveDateTime` type - requires "`chrono`" feature
-    /// Maps to `string` in TS (ISO 8601 format: "2025-11-29T14:30:00")
-    /// Zod: z.string().datetime({ local: true })
-    /// JSON Schema: string with format "date-time"
+    /// Chrono `NaiveDateTime` type - requires "`chrono`" feature.
+    /// Maps to `string` in TS (ISO 8601 format: "2025-11-29T14:30:00").
+    /// Zod: `z.string().datetime({ local: true })`.
+    /// JSON Schema: string with format "date-time".
     NaiveDateTime,
     #[cfg(feature = "chrono")]
-    /// Chrono `NaiveTime` type - requires "`chrono`" feature
-    /// Maps to `string` in TS (format: "14:30:00")
-    /// Zod: z.string().time()
-    /// JSON Schema: string with format "time"
+    /// Chrono `NaiveTime` type - requires "`chrono`" feature.
+    /// Maps to `string` in TS (format: "14:30:00").
+    /// Zod: `z.string().time()`.
+    /// JSON Schema: string with format "time".
     NaiveTime,
     #[cfg(feature = "object_id")]
-    /// `MongoDB` `ObjectId` type - requires "`object_id`" feature
-    /// Maps to `ObjectId` interface in TS with $oid: string
-    /// Zod: z.object({ $oid: z.string().regex(...) })
-    /// JSON Schema: object with $oid string property
-    /// See README.md for serialization format and validation details
+    /// `MongoDB` `ObjectId` type - requires "`object_id`" feature.
+    /// Maps to `ObjectId` interface in TS with `$oid: string`.
+    /// Zod: `z.object({ $oid: z.string().regex(...) })`.
+    /// JSON Schema: object with `$oid` string property.
+    /// See `README.md` for serialization format and validation details.
     ObjectId,
     /// Reference to another struct/enum type, potentially with generics
     /// First String is the type name (without Json suffix in TS)
@@ -356,17 +361,15 @@ impl FieldDef {
             | FieldDefType::Isize => "number".to_owned(),
             FieldDefType::F32 | FieldDefType::F64 => "number".to_owned(),
             #[cfg(feature = "object_id")]
-            FieldDefType::ObjectId => crate::features::object_id::get_object_id_typescript_type(),
+            FieldDefType::ObjectId => object_id::get_object_id_typescript_type(),
             #[cfg(feature = "chrono")]
-            FieldDefType::NaiveDate => crate::features::chrono::get_naive_date_typescript_type(),
+            FieldDefType::NaiveDate => chrono::get_naive_date_typescript_type(),
             #[cfg(feature = "chrono")]
-            FieldDefType::NaiveTime => crate::features::chrono::get_naive_time_typescript_type(),
+            FieldDefType::NaiveTime => chrono::get_naive_time_typescript_type(),
             #[cfg(feature = "chrono")]
-            FieldDefType::NaiveDateTime => {
-                crate::features::chrono::get_naive_datetime_typescript_type()
-            }
+            FieldDefType::NaiveDateTime => chrono::get_naive_datetime_typescript_type(),
             #[cfg(feature = "chrono")]
-            FieldDefType::DateTime => crate::features::chrono::get_datetime_typescript_type(),
+            FieldDefType::DateTime => chrono::get_datetime_typescript_type(),
         };
         let pre_result = if self.is_array {
             format!("Array<{result}>")
@@ -489,15 +492,15 @@ impl FieldDef {
             | FieldDefType::Isize => self.zod_number_type("z.number().int()"),
             FieldDefType::F32 | FieldDefType::F64 => self.zod_number_type("z.number()"),
             #[cfg(feature = "object_id")]
-            FieldDefType::ObjectId => crate::features::object_id::get_object_id_zod_schema(),
+            FieldDefType::ObjectId => object_id::get_object_id_zod_schema(),
             #[cfg(feature = "chrono")]
-            FieldDefType::NaiveDate => crate::features::chrono::get_naive_date_zod_schema(),
+            FieldDefType::NaiveDate => chrono::get_naive_date_zod_schema(),
             #[cfg(feature = "chrono")]
-            FieldDefType::NaiveTime => crate::features::chrono::get_naive_time_zod_schema(),
+            FieldDefType::NaiveTime => chrono::get_naive_time_zod_schema(),
             #[cfg(feature = "chrono")]
-            FieldDefType::NaiveDateTime => crate::features::chrono::get_naive_datetime_zod_schema(),
+            FieldDefType::NaiveDateTime => chrono::get_naive_datetime_zod_schema(),
             #[cfg(feature = "chrono")]
-            FieldDefType::DateTime => crate::features::chrono::get_datetime_zod_schema(),
+            FieldDefType::DateTime => chrono::get_datetime_zod_schema(),
         };
         let array_result = if self.is_array {
             format!("z.array({result})")
@@ -771,10 +774,10 @@ fn get_field_def_type_or_sibling(t_name: &str) -> FieldDefType {
         "f64" => FieldDefType::F64,
         #[cfg(feature = "object_id")]
         "ObjectId" => {
-            if crate::features::object_id::should_handle_as_object_id(t_name) {
+            if object_id::should_handle_as_object_id(t_name) {
                 FieldDefType::ObjectId
             } else {
-                FieldDefType::SiblingType(t_name.to_string(), vec![])
+                FieldDefType::SiblingType(t_name.to_owned(), vec![])
             }
         }
         #[cfg(not(feature = "object_id"))]
@@ -837,8 +840,8 @@ pub fn is_plain_enum(item_enum: &ItemEnum) -> bool {
         .all(|variant| matches!(variant.fields, Fields::Unit))
 }
 
-/// Check if a type name is a DateTime generic type (chrono feature)
-/// Returns true only when chrono feature is enabled and type is DateTime
+/// Check if a type name is a `DateTime` generic type (chrono feature).
+/// Returns true only when chrono feature is enabled and type is `DateTime`.
 #[cfg(feature = "chrono")]
 fn is_datetime_generic_type(type_name: &str) -> bool {
     type_name == "DateTime"
@@ -849,8 +852,8 @@ const fn is_datetime_generic_type(_type_name: &str) -> bool {
     false
 }
 
-/// Handle DateTime<Tz> generic type (chrono feature)
-/// Creates a FieldDef with DateTime type, ignoring the timezone parameter
+/// Handle `DateTime<Tz>` generic type (chrono feature).
+/// Creates a `FieldDef` with `DateTime` type, ignoring the timezone parameter.
 #[cfg(feature = "chrono")]
 fn handle_datetime_generic_type(safe_name: String, field_docs: &str) -> FieldDef {
     FieldDef {
@@ -859,7 +862,7 @@ fn handle_datetime_generic_type(safe_name: String, field_docs: &str) -> FieldDef
         field_type: FieldDefType::DateTime,
         is_array: false,
         array_num: None,
-        docs: field_docs.to_string(),
+        docs: field_docs.to_owned(),
         model_schema_prop_meta: None,
     }
 }
