+++ tests/chrono_tests.rs
@@ -3,6 +3,7 @@
 //! These tests verify that `NaiveDate`, `NaiveTime`, `NaiveDateTime`, and `DateTime`<Tz>
 //! types are properly converted to TypeScript types and Zod schemas.
 
-#[cfg(all(test, feature = "chrono"))]
+#[cfg(test)]
+#[cfg(feature = "chrono")]
 #[path = "chrono_tests/tests.rs"]
 mod tests;
